refactor srv_qry_mongo
