--- srv_qry_mongo/ContactsController/Program.cs
@@ -1,31 +1,0 @@
-
-var builder = WebApplication.CreateBuilder(args);
-
-// Add services to the container.
-
-builder.Services.AddControllers();
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-builder.Services.Configure<MongoDBConnect>(builder.Configuration.GetSection("ConnectionStrings"));
-builder.Services.AddSingleton<MongoDBConnect>();
-builder.Services.AddScoped<ContactsController.Controllers.MongoDBService>();
-
-
-var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseHttpsRedirection();
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
