--- srv_qry_mongo/ContactsController/appsettings.Development.json
@@ -1,11 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "ConnectionStrings": {
-    "MongoDBConnection": "mongodb://localhost:27017"
-  }
-}
