+++ srv_qry_mongo/srv_qry_mongo.Api/.gitignore
