+++ srv_qry_mongo/srv_qry_mongo.Api/appsettings.Development.json
@@ -1,0 +1,11 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "ConnectionStrings": {
+    "MongoDBConnection": "mongodb://localhost:27017"
+  }
+}
