+++ srv_qry_mongo/srv_qry_mongo.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "srv_qry_mongo", "srv_qry_mongo", "{58D866C7-55F7-4690-BB2B-CBDCFAE6C9EA}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ContactsController", "srv_qry_mongo\ContactsController\ContactsController.csproj", "{DC704891-3D6F-4452-9150-9AB0A3203499}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "srv_qry_mongo.Api", "srv_qry_mongo.Api\ContactsController.csproj", "{DC704891-3D6F-4452-9150-9AB0A3203499}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
